downloads: upgrade patchelf 0.13.1 -> 0.18.0
It's been a while since we upgraded this tool. There's a bug
fix in 0.14.0 that stops corrupting ELF binaries when we patch
DT_NEEDED entries multiple times. While we don't need it yet,
I have a commit in the works that manages to tickle the bug. So
let's take the upgrade.